feat(validation):add validation for delete owner

+++ BACK/mup-vehicles/src/main/java/com/example/mupvehicles/service/Impl/OwnerServiceImpl.java
@@ -5,9 +5,11 @@ import com.example.mupvehicles.dto.OwnerDto;
 import com.example.mupvehicles.mapper.OwnerMapper;
 import com.example.mupvehicles.model.DriverId;
 import com.example.mupvehicles.model.Owner;
+import com.example.mupvehicles.model.OwnershipTransfer;
 import com.example.mupvehicles.model.Vehicle;
 import com.example.mupvehicles.repository.DriverIdRepository;
 import com.example.mupvehicles.repository.OwnerRepository;
+import com.example.mupvehicles.repository.OwnershipTransferRepository;
 import com.example.mupvehicles.repository.VehicleRepository;
 import com.example.mupvehicles.service.OwnerService;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -24,13 +26,15 @@ public class OwnerServiceImpl implements OwnerService {
     private final VehicleRepository vehicleRepository;
     private final OwnerMapper ownerMapper;
     private final DriverIdRepository driverIdRepository;
+    private final OwnershipTransferRepository ownershipTransferRepository;
 
     @Autowired
-    public OwnerServiceImpl(OwnerRepository ownerRepository, VehicleRepository vehicleRepository, OwnerMapper ownerMapper, DriverIdRepository driverIdRepository) {
+    public OwnerServiceImpl(OwnerRepository ownerRepository, VehicleRepository vehicleRepository, OwnerMapper ownerMapper, DriverIdRepository driverIdRepository, OwnershipTransferRepository ownershipTransferRepository) {
         this.ownerRepository = ownerRepository;
         this.vehicleRepository = vehicleRepository;
         this.ownerMapper = ownerMapper;
         this.driverIdRepository = driverIdRepository;
+        this.ownershipTransferRepository = ownershipTransferRepository;
     }
 
     @Override
@@ -83,6 +87,17 @@ public class OwnerServiceImpl implements OwnerService {
         Optional<DriverId> driverId = driverIdRepository.findByOwner(owner);
         if (driverId.isPresent()) {
             throw new RuntimeException("Can not delete owner, driver has driverId");
+        }
+
+        Optional<OwnershipTransfer> ownershipTransferByNewOwner = ownershipTransferRepository.findByNewOwner(owner);
+
+        if (ownershipTransferByNewOwner.isPresent()) {
+            throw new RuntimeException("Can not delete owner, owner has ownershipTransfer");
+        }
+        Optional<OwnershipTransfer> ownershipTransferByOldOwner = ownershipTransferRepository.findByOldOwner(owner);
+
+        if (ownershipTransferByOldOwner.isPresent()) {
+            throw new RuntimeException("Can not delete owner, owner has ownershipTransfer");
         }
 
         ownerRepository.delete(owner);
